requirements.txt
pyTelegramBotAPI==4.14.0
kucoin-python==1.0.26
pandas==2.0.3
numpy==1.24.3
requests==2.31.0